+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes to the `dply` crate are documented in this file.
 * `parquet`: A folder path reads all parquet files in the folder.
 * `parquet`: write now uses compression to reduce file size.
 * `anti_join`: Added `anti_join` to select rows not found in the orther data frame.
+* `mutate`: Make `len` function to work on strings.
 
 ## 0.1.9 - 2023-06-28
 ### 🔧 Changed